+++ frontend/src/app/components/staff-login/staff-login.component.ts
@@ -18,7 +18,7 @@ export class StaffLoginComponent {
     password = '';
     code = '';
   
-    loginAdmin() {
+    loginStaff() {
       this.loginService.loginStaff(this.username, this.password).subscribe({
         next: (adminInfo: StaffResponseDTO) => {
           console.log('Login successful:', adminInfo);