+++ frontend/CHANGELOG.md
@@ -5,8 +5,14 @@
 - Test of changeLog
 - Creation of Login and Welcome Page
 
-## [v0.0.1] - 2025-05-27
+## [v0.0.1] - 2025-05-27 
 ### WIP Version
 - Admin menus
+
+## [v0.0.2] - 2025-05-30 
+### WIP Version
+- Admin staff modify is created
+- Staff login works
+